Set the grub2 timeout to 10 seconds in the base role
In some cases, we need to access the console of one of our VMs to
do some maintenance (e.g. if we get a kernel panic). For those cases,
the default grub2 timeout of 1 second makes it very difficult to
stop the boot process.

Increasing the timeout to 10 seconds should give us enough room.

Change-Id: I2441b19738d18d278ca562b6f7986a7543f154ef