Manage the Sensu repository ourselves
When manage_repos is true, opstools-ansible wants to install a lot of
repositories such as centos-release-openstack-* and the likelihood
of those repositories conflicting with the software meant to run on
our servers is pretty high.

Therefore, disable the opstools repository management and only install
the sensu repository -- that's all we need.

Change-Id: I05f953d8d6e3579af3f067be75f266c7db5395a6