Add monitoring for Nodepool image age
In some cases, one of the nodepool images can fail to build for
several reasons. When this happens, it can be a potential issue,
specially for the upstream-centos-7 image. Adding a script to
monitor the age of a nodepool image, and monitoring for the current
images in review.rdoproject.org.

Change-Id: Id4aa267e8daad4109a224b124aeae2c19174e5f4