Initial operations/troubleshooting doc for RDO Infra
We want to create some troubleshooting and operations documentation
for RDO Infra services, to spread the knowledge internally and avoid
the bus factor [1].

This is an initial commit to create the basic structure, and adds
documentation for some components (RDO Trunk servers and Nodepool).
Placeholders are in place for other services.

[1] - https://en.wikipedia.org/wiki/Bus_factor

Change-Id: I962f52de7dcef0b50af2503bb8ecea6db030951c