Fixes to registry pruning script
If, for some reason, the openshift_tag_whitelist.sh fails and
provides an empty whitelist, we could remove some tags that are
required. We do not want that to happen when we are running the
job with cron, so fail when the whitelist is empty.

Also, add output to logfile for the main script.

Change-Id: I116408d6d2904972b01260b24fe2ab151429bfc0