[WIP] Allow removing users from the system
We want to be able to remove users from our systems, if they move
to other duties. This change allows us to specify that an admin
user should be removed, by adding a state: 'absent' line, like the
following:

   jpena:
     name: "jpena"
     state: "absent"
     ssh_keys:
       - "ssh-rsa A..."

If no state is specified, it is assumed to be "present".

TG-1681

Change-Id: Ie887a22d11351b6931595e0fbe30356e7559a9b1