Add playbooks to setup the RDO Trunk replicated DB
We already had documented the manual steps to create the RDO Trunk
database and its replica in [1], but it is always better to have a
playbook that does the job for you.

This commit adds the dlrn-replicated-db.yml, which will create a
replicated database configuration. Individual databases for each
DLRN instance are not covered here.

[1] - https://github.com/rdo-infra/rdo-infra-playbooks/blob/master/doc/source/dbreplica.rst
TG-1729

Change-Id: Ia95ccf83903bd8fa7bbc088879dfab43a8b92cee