Add check to verify last sync of centos AFS mirror
We've had a case where the openstack-infra AFS mirror for centos.org
was outdated, and CI jobs were failing because of that. This commit
adds a new check, that uses the openstack infra Graphite data to
alert us if the mirror has not been updated in the last 24 hours. It
takes the raw data used by [1].

[1] - http://grafana.openstack.org/d/ACtl1JSmz/afs?orgId=1&panelId=14&fullscreen

Change-Id: Icdf020f218efa94e2e8614e4fcda20b7118efe3a