Toggle off HSTS for www.rdoproject.org
The force_ssl variable doesn't seem to have ever existed in the
httpd role.
We need to toggle off HSTS for rdoproject.org in order to allow
subdomains without SSL while we work on adding SSL to them.

This is safe to merge but will need a patch to land in the role
first [1] in order to be effective.

[1]: https://gitlab.com/osas/ansible-role-ah-httpd/merge_requests/58/diffs?commit_id=eb50ce36bb8f015d3bd0de0c5035247bcf97af72

Change-Id: I9f6a516ed276d25b6605bbb4017238a605801fe3