Adapt image-server role to per-distro directories
Since we are going to have Fedora 28-based images (and in the future
we can expect CentOS 8-based ones), we are going to set up a new
structure for images.rdoproject.org.

The base directory will now be:
  /var/www/html/<distro>/<release>/rdo_trunk

We will be keeping backwards compatibility with URLs, so centos7/master
will symlink to /master and so on.

This also requires changes in the image pruning script.

Change-Id: Ibb6102bb53cda0e4e23577044cda4bcf4ee8bbec