Update patch set 1
Patch Set 1:

There is a serious issue with use of HSTS on bare domain name: as it makes almost impossible to use any non sercure HTTP service on any of the subdomains. 

HSTS is caches for ever/years by the browsers so if you do not want to spend enless hours asking people to remove cache (which I hate!) we are better off without it. Using it for non-root domains is perfectly fine, only root ones are the big trouble.

Remember that there are still valid use cases for plain HTTP, as this can be cached opposed to HTTPS which cannot practically be cached.

Patch-set: 1