Use OpenShift Origin 3.11 for registry
This commit adds support to deploy the RDO Registry based on
Origin 3.11. It comes from the previous work in [1], and has been
adapted to the new playbook structure.

Note that, starting with this version, we will no longer be deploying
a fully functional web console, as it has not been used so far and
requires maintaining a custom patch to the openshift-ansible roles.

[1] - https://review.rdoproject.org/r/14181

Change-Id: I88cad8a70a8a7835e6bd469d35578319baa14cb4
Co-Authored-By: David Moreau Simard <dms@redhat.com>