Update patch set 1
Patch Set 1:

It is not a bug, it is a wanted feature.

If there are valid use cases for unsecure web, I'd like to know because I don't see any. I mean it.

In the past it was complicated to get certificates but it is now just totally trivial and you can have your site working in a matter of seconds, so I don't see where the burden would be.

Experience shows people install "test" sites, or are in a hurry and deffer installing certificates, and then the sites' auth is plugged into the usual auth system for convenience, or it becomes production but noone cares to make it secure before, and so on and so on. These are real problems we see just everyday.

So it's not a problem of technical implementation, it's a problem of disagreement on the existence of such insecure sites.

Patch-set: 1