Update patch set 1
Patch Set 1:

I think that Alan nailed the best explanation: "HSTS on by default is fine for explicitly configured websites".

The issue comes from bare-domains using it with includesubdomain, which effectively prevents any subdomain from working in non SSL.

This can prove to be a problem with mirrors as it prevents others from implementing locally caching-proxies/mirrors if the upstream site is on HTTPS.

Maybe the title is a big confusing because I do not see any issue on having HSTS enabled on www. domain, but when it comes to non www. prefixed one, we can do it but only without includesubdomain.

Patch-set: 1