Update patch set 1
Patch Set 1:

It seems you copied the configuration from "Intermediate (OpenSSH 5.3): but EL7 has a better version (7.4) so "Modern (OpenSSH 6.7+)" would be a better fit. Currently you're excluding recommended and supported algos like 'curve25519' (according to the same document, which I believe is a good one too).

As for Infosec's recommendation, I already knew about it but I'm discussing with them for a better option. Their main goal is to disable weak algorithms but the openssh package we ship as a distro diverge from upstream and reenables some algorithms, so my hope is having this fixed in the packaging as it affects everybody using our distros. I would then recommend to defer such changes.

In the future it should be easier because options allow to only disable algos without redefining the whole list.

As for PermitRootLogin I see your point but this means setting it to "No" (which concurs with the same document) but you simply removed the Ansible rule so it will not be enforced.

Because of all the above I still disagree with this PR as it is.

Patch-set: 1