Change rcm symlink owner; changed rhel dir permissions
Currently ansible file module when its creating symlink with
setting owner to a dir, it sets owner on both location.
In that case, setting permissions to 750 to the rcm content
dir not possible using the file module.
Set rcm_share user as a owner/group for that dir and remove access to
others (o-rwx)

Change-Id: I3c7518c136d55f159758dfa8b22e09aa03932deb