Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Are we planning to build container images based on the current CentOS Stream 9 packages? 
> 

Yes Correct so that we are ready before CentOS stream 9 hits our door.

> Dependencies have not been built in Koji yet, and the DLRN API endpoint for the centos9-master-preview worker is not really functional at the moment.

Currently we are not aiming to promote any content. We will be adding the jobs and running via testproject.

We need a namespace on rdoregistry to push the built containers so that we can consume it.

Patch-set: 1